+++ app/ExternalFiles/Synchronization.php
@@ -434,6 +434,7 @@ class Synchronization {
 
 		// add counter handling.
 		add_action( 'eml_after_file_save', array( $this, 'update_url_count' ), 10, 0 );
+		add_action( 'eml_file_directory_import_files', array( $this, 'set_url_max_count' ), 10, 2 );
 		add_action( 'eml_ftp_directory_import_files', array( $this, 'set_url_max_count' ), 10, 2 );
 		add_action( 'eml_http_directory_import_files', array( $this, 'set_url_max_count' ), 10, 2 );
 		add_action( 'eml_sftp_directory_import_files', array( $this, 'set_url_max_count' ), 10, 2 );