+++ changelog.md
@@ -22,6 +22,7 @@
 - Added check for PHP strict usage on every release with PHPStan
 - Added support for custom Download Lists of the plugin "Download Lists with icons" incl. sync of them
 - Added support for plugin Folders to import external files in its folders
+- Added support for plugin "Media Library Organizer" to import external files in its categories
 - Added new file type "PDF" and "ZIP" for better supporting the handling of these files
 - Added info about external files in attachment modal